requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
nodeenv==1.9.1
openai==1.58.1
pdf2image==1.17.0
pillow==11.0.0
prisma==0.15.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
StrEnum==0.4.15
tiktoken==0.8.0
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
